REFACTOR :: 2차 전형 점수 입력 코드
